Add files via upload
we conducted a research project to identify and understand how different demographic attributes are associated with risk propensity of people. The research method is the application of data analytics techniques and tools including statistical scoring, normality tests, and comparison of means. The source data was a survey conducted in an earlier research by 12 conductors and with 2,278 respondents from different educational and work backgrounds, as well as different universities. Gender, English level, and German level were found to be attributes that result in statistically significant differences in risk propensity scores. However, it was also found that the conductor of the survey may also be affecting the scores. Specifically in our analysis, we found that one conductor had much lower scores than others. So, as future research, the analysis can be conducted again, excluding the 237 survey responses & data points from that outlier conductor.